end CRUD Apresentação

+++ admin/company/delete_company.php
@@ -5,13 +5,22 @@
 
         $id = $_GET['id'];
 
-        $querySelect = "SELECT *  FROM company WHERE companyId = $id";
+        $query = "SELECT * FROM company WHERE companyId = $id";
+        $result = $conexao->query($query);
+        
+        
+        $queryUser = "SELECT * FROM users WHERE company_id = $id";
+        $resultUser = $conexao->query($queryUser);
+        print_r($resultUser);
 
-        $result = $conexao->query($querySelect);
 
-        if($result->num_rows > 0){
+        if($result->num_rows > 0 && $resultUser->num_rows < 1){
             $queryDelete = "DELETE FROM company WHERE companyId = '{$id}'";
             $resultDelete = $conexao->query($queryDelete);
+
+        }else{
+            header('Location: ../tabelaEmpresas.php');
+            exit();         
         }
     }
     header('Location: ../tabelaEmpresas.php');
